移動処理修正

+++ 2dAction/mainLoop.hpp
@@ -14,9 +14,9 @@ private:
 	SDL_Texture *m_loadingText;//ロード画面用文字
 	vector<vector<ENEMY_ACTION>> m_enemyAction;//各敵が持つアクション
 	vector<vector<vector<int>>> m_maxFrame;//各アニメーションのフレーム数
-	vector<vector<unsigned short>> m_attackActive;//攻撃判定のあるフレーム
+	vector<vector<unsigned int>> m_attackActive;//攻撃判定のあるフレーム
 	vector<vector<vector<vector<int>>>> m_attackRect;//攻撃判定範囲
-	vector<vector<unsigned short>> m_hurtActive;//くらい判定のあるフレーム
+	vector<vector<unsigned int>> m_hurtActive;//くらい判定のあるフレーム
 	vector<vector<vector<vector<int>>>> m_hurtRect;//くらい判定範囲
 
 public:
